ci: make test.sh repeatable

+++ test/test.sh
@@ -2,22 +2,23 @@
 if which karina; then
   BIN=$(which karina)
 else
-  BIN=
   BIN=./.bin/karina
   chmod +x $BIN
   mkdir -p .bin
 fi
-export GO_VERSION=${GO_VERSION:-1.13}
 export KUBECONFIG=~/.kube/config
 REPO=$(basename $(git remote get-url origin | sed 's/\.git//'))
 GITHUB_OWNER=$(basename $(dirname $(git remote get-url origin | sed 's/\.git//')))
 GITHUB_OWNER=${GITHUB_OWNER##*:}
 MASTER_HEAD=$(curl -s https://api.github.com/repos/$GITHUB_OWNER/$REPO/commits/master | jq -r '.sha')
 
-[ -z "$KUBERNETES_VERSION" ] && echo -e "KUBERNETES_VERSION not set! Try: \nexport KUBERNETES_VERSION='v1.16.9'" && exit 1
-[ -z "$SUITE" ] && echo -e "SUITE not set! Try: \n export SUITE='minimal'\n or one of these (minus extension):"&& ls  test/*.yaml && exit 1
-kind delete cluster --name kind-$SUITE-$KUBERNETES_VERSION || echo "No cluster present when starting"
+export KUBERNETES_VERSION=${KUBERNETES_VERSION:-v1.18.6}
+export SUITE=${SUITE:-minimal}
+export CLUSTER_NAME=kind-$SUITE-$KUBERNETES_VERSION
 
+if [[ "$CI" == "true" ]]; then
+  kind delete cluster --name  $CLUSTER_NAME|| echo "No cluster present when starting"
+fi
 export CONFIGURED_VALUE=$(openssl rand -base64 12)
 export PLATFORM_CONFIG=test/$SUITE.yaml
 echo "::endgroup::"
@@ -27,15 +28,18 @@ function report() {
   set +e
   echo "::group::Uploading test results"
   if [[ "$CI" == "true" ]]; then
-    wget -nv -nc -O build-tools \
-      https://github.com/flanksource/build-tools/releases/latest/download/build-tools && \
-      chmod +x build-tools
 
-    ./build-tools junit gh-workflow-commands test-results/results.xml
+    if [[ -e test-results/results.xml ]]; then
+      wget -nv -nc -O build-tools \
+        https://github.com/flanksource/build-tools/releases/latest/download/build-tools && \
+        chmod +x build-tools
 
+      ./build-tools junit gh-workflow-commands test-results/results.xml
+    fi
     mkdir -p artifacts
-    $BIN snapshot --output-dir snapshot -v --include-specs=true --include-logs=true --include-events=true || echo "::error::Error while creating snapshot"
-    zip -r artifacts/snapshot.zip snapshot/* || echo "::error::Error while zipping snapshot"
+    if $BIN snapshot --output-dir snapshot -v --include-specs=true --include-logs=true --include-events=true ; then
+      zip -r artifacts/snapshot.zip snapshot/*
+    fi
   else
     echo "Skipping test report when not running in CI"
   fi
@@ -44,12 +48,15 @@ function report() {
 trap report EXIT
 
 set -e
+echo "$(kubectl config current-context) != kind-$CLUSTER_NAME"
 
-if [[ "$KUBECONFIG" != "$HOME/.kube/kind-config-kind" ]] ; then
+if [[ "$(kubectl config current-context)" != "kind-$CLUSTER_NAME" ]] ; then
   echo "::group::Provisioning"
-  $BIN ca generate --name root-ca --cert-path .certs/root-ca.crt --private-key-path .certs/root-ca.key --password foobar  --expiry 1
-  $BIN ca generate --name ingress-ca --cert-path .certs/ingress-ca.crt --private-key-path .certs/ingress-ca.key --password foobar  --expiry 1
-  $BIN ca generate --name sealed-secrets --cert-path .certs/sealed-secrets-crt.pem --private-key-path .certs/sealed-secrets-key.pem --password foobar  --expiry 1
+  if [[ ! -e .certs/root-ca.key ]]; then
+    $BIN ca generate --name root-ca --cert-path .certs/root-ca.crt --private-key-path .certs/root-ca.key --password foobar  --expiry 1
+    $BIN ca generate --name ingress-ca --cert-path .certs/ingress-ca.crt --private-key-path .certs/ingress-ca.key --password foobar  --expiry 1
+    $BIN ca generate --name sealed-secrets --cert-path .certs/sealed-secrets-crt.pem --private-key-path .certs/sealed-secrets-key.pem --password foobar  --expiry 1
+  fi
   if $BIN provision kind-cluster --trace -vv ; then
     echo "::endgroup::"
   else
@@ -57,8 +64,6 @@ if [[ "$KUBECONFIG" != "$HOME/.kube/kind-config-kind" ]] ; then
     exit 1
   fi
 fi
-
-
 
 $BIN version
